requirements.txt
# System Requirement: sudo apt-get install -y libpcap-dev
paho-mqtt==2.1.0
scapy==2.5.0
pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
matplotlib==3.8.4
argparse
